Create Job Offer email template
GoPhish is great and I'm going to deploy in my org.  If this were to gain a robust template and landing page repo, it would be even better.  Here's a simple one I simply imported from my spam box.  Should be no issues with copyrighting, etc. If we can get the almighty internet to start contributing, might fill up fast.